requirements.txt
google-generativeai
python-dotenv
pillow
requests
streamlit
PyMuPDF
fpdf
